drivers/auxdisplay: sensehat: Raspberry Pi Sense HAT display driver
This patch adds the driver for the 8x8 RGB LED matrix display
on the Sense HAT. It appears as a character device named sense-hat
in /dev/. That special file is 192 bytes large and contains one byte
for each color channel of each pixel. The overall layout is:

    col 1       col 8
      |           |
      v           v
0x00: R   . . .   R \
0x08: G   . . .   G |> row 1
0x10: B   . . .   B /
      .   .       .
...   .     .     .
      .       .   .
0xA8: R   . . .   R \
0xB0: G   . . .   G |> row 8
0xB8: G   . . .   G /

Each channel may have any value between 0 and 31 (larger values are
wrapped) and are translated by a gamma table before being send to the
device to provide a linear experience of brightness (the gamma table
can be modified or reset using an ioctl call on the file). The constants
for the ioctl are in the sensehat.h header also added in this patch.

Co-developed-by: Mwesigwa Guma <mguma@redhat.com>
Signed-off-by: Mwesigwa Guma <mguma@redhat.com>
Co-developed-by: Joel Savitz <jsavitz@redhat.com>
Signed-off-by: Joel Savitz <jsavitz@redhat.com>
Co-developed-by: Daniel Bauman <dbauman@redhat.com>
Signed-off-by: Daniel Bauman <dbauman@redhat.com>
Signed-off-by: Charles Mirabile <cmirabil@redhat.com>